Move implementation out of asio namespace/directory
This implementation is not part of boost::asio so move it out of the
boost::asio namespace and directory.

It is intended to be a boost library, at least for now by name, so
keep it in the boost namespace and directory.
