--- include/boost/asio/windows_sspi.hpp
@@ -1,19 +1,0 @@
-//
-// windows_sspi.hpp
-// ~~~~~~~~~~~~~~~~
-//
-// Copyright (c) 2020 Kasper Laudrup (laudrup at stacktrace dot dk)
-//
-// Distributed under the Boost Software License, Version 1.0. (See accompanying
-// file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
-//
-
-#ifndef BOOST_ASIO_WINDOWS_SSPI_HPP
-#define BOOST_ASIO_WINDOWS_SSPI_HPP
-
-#include <boost/asio/windows_sspi/context.hpp>
-#include <boost/asio/windows_sspi/error.hpp>
-#include <boost/asio/windows_sspi/stream.hpp>
-#include <boost/asio/windows_sspi/stream_base.hpp>
-
-#endif // BOOST_ASIO_WINDOWS_SSPI_HPP
