+++ include/boost/windows_sspi/context.hpp
@@ -8,12 +8,12 @@
 // file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 //
 
-#ifndef BOOST_ASIO_WINDOWS_SSPI_CONTEXT_HPP
-#define BOOST_ASIO_WINDOWS_SSPI_CONTEXT_HPP
+#ifndef BOOST_WINDOWS_SSPI_CONTEXT_HPP
+#define BOOST_WINDOWS_SSPI_CONTEXT_HPP
 
-#include <boost/asio/windows_sspi/context_base.hpp>
-#include <boost/asio/windows_sspi/detail/sspi_functions.hpp>
-#include <boost/asio/windows_sspi/error.hpp>
+#include <boost/windows_sspi/context_base.hpp>
+#include <boost/windows_sspi/detail/sspi_functions.hpp>
+#include <boost/windows_sspi/error.hpp>
 
 // TODO: Avoid cluttering global namespace (and avoid Windows headers if possible)
 #define SECURITY_WIN32
@@ -24,7 +24,6 @@
 #include <stdexcept>
 
 namespace boost {
-namespace asio {
 namespace windows_sspi {
 
 class context : public context_base {
@@ -73,7 +72,6 @@ private:
 };
 
 } // namespace windows_sspi
-} // namespace asio
 } // namespace boost
 
-#endif // BOOST_ASIO_WINDOWS_SSPI_CONTEXT_HPP
+#endif // BOOST_WINDOWS_SSPI_CONTEXT_HPP
