+++ include/boost/windows_sspi/context_base.hpp
@@ -8,11 +8,10 @@
 // file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 //
 
-#ifndef BOOST_ASIO_WINDOWS_SSPI_CONTEXT_BASE_HPP
-#define BOOST_ASIO_WINDOWS_SSPI_CONTEXT_BASE_HPP
+#ifndef BOOST_WINDOWS_SSPI_CONTEXT_BASE_HPP
+#define BOOST_WINDOWS_SSPI_CONTEXT_BASE_HPP
 
 namespace boost {
-namespace asio {
 namespace windows_sspi {
 
 class context_base {
@@ -95,7 +94,6 @@ public:
 };
 
 } // namespace windows_sspi
-} // namespace asio
 } // namespace boost
 
-#endif // BOOST_ASIO_WINDOWS_SSPI_CONTEXT_BASE_HPP
+#endif // BOOST_WINDOWS_SSPI_CONTEXT_BASE_HPP
