+++ include/boost/windows_sspi/windows_sspi.hpp
@@ -1,0 +1,19 @@
+//
+// windows_sspi.hpp
+// ~~~~~~~~~~~~~~~~
+//
+// Copyright (c) 2020 Kasper Laudrup (laudrup at stacktrace dot dk)
+//
+// Distributed under the Boost Software License, Version 1.0. (See accompanying
+// file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+//
+
+#ifndef BOOST_WINDOWS_SSPI_HPP
+#define BOOST_WINDOWS_SSPI_HPP
+
+#include <boost/windows_sspi/context.hpp>
+#include <boost/windows_sspi/error.hpp>
+#include <boost/windows_sspi/stream.hpp>
+#include <boost/windows_sspi/stream_base.hpp>
+
+#endif // BOOST_WINDOWS_SSPI_HPP
